+++ hockey_scraper/nhl/json_schedule.py
@@ -41,7 +41,7 @@ def chunk_schedule_calls(from_date, to_date):
     :return: raw json of schedule of date range
     """
     sched = []
-    days_per_call = 50
+    days_per_call = 30
 
     from_date = datetime.strptime(from_date, "%Y-%m-%d") 
     to_date = datetime.strptime(to_date, "%Y-%m-%d")